# backend/openapi/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "GifJam.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/games": {
      "post": {
        "tags": [
          "Games"
        ],
        "operationId": "CreateGame",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerGameSnapshot"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/games/{code}/join": {
      "post": {
        "tags": [
          "Games"
        ],
        "operationId": "JoinGame",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerGameSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/games/{code}/settings": {
      "put": {
        "tags": [
          "Games"
        ],
        "operationId": "UpdateGameSettings",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LobbySnapshot"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/games/{code}/leave": {
      "post": {
        "tags": [
          "Games"
        ],
        "operationId": "LeaveGame",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/games/{code}": {
      "get": {
        "tags": [
          "Games"
        ],
        "operationId": "GetGame",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerGameSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/matchmaking/join": {
      "post": {
        "tags": [
          "Matchmaking"
        ],
        "operationId": "JoinMatchmaking",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchmakingSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/matchmaking/leave": {
      "post": {
        "tags": [
          "Matchmaking"
        ],
        "operationId": "LeaveMatchmaking",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/matchmaking/status": {
      "get": {
        "tags": [
          "Matchmaking"
        ],
        "operationId": "GetMatchmakingStatus",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchmakingSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateGameRequest": {
        "type": "object",
        "properties": {
          "totalRounds": {
            "type": "integer",
            "format": "int32"
          },
          "phraseSubmissionSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "resultsSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "mode": {
            "$ref": "#/components/schemas/GameMode"
          }
        },
        "additionalProperties": false
      },
      "GameMode": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "GameStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "LobbyPlayerSnapshot": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "isReady": {
            "type": "boolean"
          },
          "isConnected": {
            "type": "boolean"
          },
          "isHost": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "LobbySnapshot": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/GameStatus"
          },
          "mode": {
            "$ref": "#/components/schemas/GameMode"
          },
          "totalRounds": {
            "type": "integer",
            "format": "int32"
          },
          "phraseSubmissionSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "resultsSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "currentRoundNumber": {
            "type": "integer",
            "format": "int32"
          },
          "hostUserId": {
            "type": "string",
            "format": "uuid"
          },
          "canStart": {
            "type": "boolean"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LobbyPlayerSnapshot"
            },
            "nullable": true
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "MatchmakingSnapshot": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/MatchmakingStatus"
          },
          "playerCount": {
            "type": "integer",
            "format": "int32"
          },
          "minimumPlayers": {
            "type": "integer",
            "format": "int32"
          },
          "maximumPlayers": {
            "type": "integer",
            "format": "int32"
          },
          "hostUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "deadlineAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "gameCode": {
            "type": "string",
            "nullable": true
          },
          "gameMode": {
            "$ref": "#/components/schemas/GameMode"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "MatchmakingStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "PhraseSource": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "PlayerGameSnapshot": {
        "type": "object",
        "properties": {
          "lobby": {
            "$ref": "#/components/schemas/LobbySnapshot"
          },
          "isHost": {
            "type": "boolean"
          },
          "round": {
            "$ref": "#/components/schemas/PlayerRoundSnapshot"
          }
        },
        "additionalProperties": false
      },
      "PlayerGifSelectionSnapshot": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          },
          "mediaUrl": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "previewWidth": {
            "type": "integer",
            "format": "int32"
          },
          "previewHeight": {
            "type": "integer",
            "format": "int32"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true
          },
          "attribution": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerGifSnapshot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          },
          "mediaUrl": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "previewWidth": {
            "type": "integer",
            "format": "int32"
          },
          "previewHeight": {
            "type": "integer",
            "format": "int32"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true
          },
          "attribution": {
            "type": "string",
            "nullable": true
          },
          "isOwn": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "PlayerPhraseSnapshot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "isOwn": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "PlayerRoundSnapshot": {
        "type": "object",
        "properties": {
          "roundNumber": {
            "type": "integer",
            "format": "int32"
          },
          "phase": {
            "$ref": "#/components/schemas/RoundPhase"
          },
          "phaseEndsAt": {
            "type": "string",
            "format": "date-time"
          },
          "gifVotingPresentationEndsAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "hasSubmittedPhrase": {
            "type": "boolean"
          },
          "hasVotedPhrase": {
            "type": "boolean"
          },
          "hasSubmittedGif": {
            "type": "boolean"
          },
          "hasVotedGif": {
            "type": "boolean"
          },
          "hasConfirmedResults": {
            "type": "boolean"
          },
          "phrases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerPhraseSnapshot"
            },
            "nullable": true
          },
          "gifs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerGifSnapshot"
            },
            "nullable": true
          },
          "selectedPhrase": {
            "$ref": "#/components/schemas/SelectedPhraseSnapshot"
          },
          "gifSelection": {
            "$ref": "#/components/schemas/PlayerGifSelectionSnapshot"
          },
          "reveal": {
            "$ref": "#/components/schemas/RoundRevealSnapshot"
          },
          "ranking": {
            "$ref": "#/components/schemas/RankingSnapshot"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RankingEntrySnapshot": {
        "type": "object",
        "properties": {
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RankingSnapshot": {
        "type": "object",
        "properties": {
          "gameCode": {
            "type": "string",
            "nullable": true
          },
          "completedRounds": {
            "type": "integer",
            "format": "int32"
          },
          "isFinal": {
            "type": "boolean"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RankingEntrySnapshot"
            },
            "nullable": true
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "RevealedGifSnapshot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          },
          "mediaUrl": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "previewWidth": {
            "type": "integer",
            "format": "int32"
          },
          "previewHeight": {
            "type": "integer",
            "format": "int32"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true
          },
          "attribution": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/RevealedPlayerSnapshot"
          },
          "voteCount": {
            "type": "integer",
            "format": "int32"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RevealedPhraseSnapshot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/PhraseSource"
          },
          "author": {
            "$ref": "#/components/schemas/RevealedPlayerSnapshot"
          }
        },
        "additionalProperties": false
      },
      "RevealedPlayerSnapshot": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RoundPhase": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "RoundRevealSnapshot": {
        "type": "object",
        "properties": {
          "roundNumber": {
            "type": "integer",
            "format": "int32"
          },
          "phrase": {
            "$ref": "#/components/schemas/RevealedPhraseSnapshot"
          },
          "gifs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RevealedGifSnapshot"
            },
            "nullable": true
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "SelectedPhraseSnapshot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateGameSettingsRequest": {
        "type": "object",
        "properties": {
          "totalRounds": {
            "type": "integer",
            "format": "int32"
          },
          "phraseSubmissionSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "resultsSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "mode": {
            "$ref": "#/components/schemas/GameMode"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}